requirements.txt
alembic==1.13.3
beautifulsoup4==4.12.3
cohere==5.13.4
faiss_cpu==1.8.0
fitz==0.0.1.dev2
Flask==3.1.0
Flask_Cors==4.0.1
Flask_Migrate==4.0.7
Flask_SocketIO==5.3.6
flask_sqlalchemy==3.1.1
google_api_python_client==2.149.0
google_auth_oauthlib==1.2.1
groq==0.18.0
gunicorn==21.2.0
huggingface_hub==0.26.3
nltk==3.8.1
numpy==1.23.5
protobuf==3.20.3
psycopg2-binary==2.9.8
pusher==3.3.2
pymongo==4.10.1
pymupdf==1.25.1
python-dotenv==1.0.1
rank_bm25==0.2.2
Requests==2.32.3
scikit_learn==1.1.3
sentence_transformers==3.0.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.32
torch==2.5.1
transformers==4.46.3
Werkzeug==3.1.3
